Fix the new tool lockfile deprecation message. (#19565)
When displaying the list of requirements you need in your new tool
lockfile, it called `self.requirements`, which used to just be a helper
function to return the old-style requirements. But at some point it
became an option specifying the new-style requirements, and this
callsite wasn't updated. It is now.