Add ruff, adjust pyugprade, autoflake in docs (#19555)
Fixes #19273 by adding ruff to the docs, both the auto-generated backend
docs and the manual pages that list out backends. This doesn't try to
solve the general problem of reminding us to add backends to docs.

While I was there, I noticed the pyupgrade and autoflake backends were
referring to the `experimental` paths, but they've now graduated to
non-experimental, so this switches them over.