upgrade tonic ecosystem (gRPC crates) to latest versions (#19517)
Upgrade the Tonic ecosystem crates:
- `tonic` v0.9.x
- `prost` v0.11.x

Notable changes:
- `grpc_util::create_endpoint` is now `async` and some of the Pants
bootstrap logic must become `async` higher in the call stack to
accommodate that. It is also renamed to `grpc_util::create_channel`.
- Tonic has relaxed the `Sync` requirement on services; most of them (if
not all) can be just `Sync` now instead of `Send + Sync`.
- Introduce `grpc_util::channel::Channel` to represent a combined HTTP /
HTTPS connector. This approach is inspired by
https://github.com/LucioFranco/tonic-openssl/blob/master/example/src/client2.rs.
We could conceivably also add OpenSSL support in the future by extending
the `grpc_util::channel::Client` enumeration to be based on the
`tonic-openssl` crate.
- CI: The `prost` crate no longer bundles the `protoc` protobuf
compiler. We use the `arduino/setup-protoc` action to install it. The
repository token is passed to the action in order to avoid rate limits
in downloading `protoc`. (Hitting GitHub rate limits was actually an
issue during testing of this PR.)

---------

Co-authored-by: Tom Dyas <tdyas@toolchain.com>