Don't override the ICs in the pants script on MacOS Silicon.
We used to need this because pants ran on Python 3.7-3.9
generally, but only on 3.9 on MacOS Silicon.

It is no longer necessary to do so, since the repo-wide
ICs are 3.9.* anyway now.

Setting this was not harmless, as it would override the ICs in
other repos when running from sources with PANTS_SOURCE=.