Sunset uploading to binaries.pantsbuild.org (#19564)
This change comes as a result of all of the efforts to switch our
distribution model away from PyPI/S3 and into GitHub. It's the first
change that takes away instead of adds. However, now that we're fully on
GitHub Releases, we can wean off S3/PyPI.

Specifically this change:
- Changes our release GitHub Action:
  - Uploads wheels inside each build job, instead of uploading to S3
- Downloads the wheels from the GitHub Release to push to PyPI, instead
of downloading from S3 (and reversioning)
- Changes our release script:
  - Moved `build-wheels` to do the `manylinux` renaming
  - A few refactorings to eliminate SHAs from versions/paths
  - Removed a whole lotta now-dead code (more to come :wink: )
- Removed other dead bits