Show cause when ./pants can't find interpreter (#19567)
Surfaces an existing error message when a pants developer runs the
`./pants` bootstrap script and it fails because it can't find a Python
3.9 interpreter. (Which was [happening to
me](https://pantsbuild.slack.com/archives/C046T6T9U/p1691145625985819))

## before
```bash
~/code/pantsbuild/pants (main) $ command -v python3.9
[Exit:1] ~/code/pantsbuild/pants (main) $ ./pants
[Exit:1] ~/code/pantsbuild/pants (main) $
```
./pants fails with no indication of why.

## after
```bash
~/code/pantsbuild/pants (for-devs-err-msg-if-no-python3.9) $ ./pants
pants: Failed to find a Python 3.9 interpreter
[Exit:1] ~/code/pantsbuild/pants (for-devs-err-msg-if-no-python3.9) $ 
```

## implementation
In main, we end up swallowing the error because we call
`determine_python` with
```bash
# ./pants shell script
# ...
PY="$(determine_python)"
```

`determine_python` returns with 1 so the script fails but it emitted the
error message to stdout which was redirected away from the console by
`$(...)`. The PR emits the error to stderr instead so it shows up.

Co-authored-by: Alexey Tereshenkov <50622389+AlexTereshenkov@users.noreply.github.com>