Make mypy cache move truly atomic (Cherry-pick of #19610) (#19615)
This change makes the post-mypy-run "moving" of the mypy cache database
truly atomic (as claimed in the comment).

It does so by checking the device IDs of the source and dest. If they
match, `mv` is atomic. Otherwise, `cp` the src into a tmpfile next to
the dest, then `mv` into place.

Fixes https://github.com/pantsbuild/pants/issues/19609

Co-authored-by: Josh Cannon <joshdcannon@gmail.com>