Require scie-pants 0.9.2 or newer, for new distribution model (#19654)
This updates Pants to require and check for the scie-pants launcher
binary version: we need version 0.9 or earlier, to have support for the
new distribution model. This syncs with
https://github.com/pantsbuild/scie-pants/pull/246 (which will become
scie-pants 0.9.2 / 0.9.3), which sets the `SCIE_PANTS_VERSION`
environment variable when running pants.

This will be cherry-picked back as far as we support, to help users
upgrade their launcher earlier.

Fixes #19600