Add lint backend to run semgrep (#18593)
This adds an `pants.backend.experimental.tools.semgrep` backend that
runs the Semgrep (https://semgrep.dev) OSS engine
(https://semgrep.dev/docs/getting-started/) as a linter.

### Background

Semgrep is a multilanguage linter, that allows defining in-repo
per-language rules similar to a PyLint plugin for Python, but a
source-like syntax for defining the rules (rather than manual AST
traversal). It has explicit support for [many
languages](https://semgrep.dev/docs/supported-languages/), and even a
`generic` language that works on any (text) file.

We use this for a whole bunch of checks like:

- every GitHub Actions job has a timeout configured in the workflow YAML
file
- erroring if we call `logging.info(...)` in Python, rather than using
the appropriate method on a `logging.Logger` instance

A random open source example:
https://github.com/aiven/terraform-provider-aiven/blob/ae42e99db2ee1e56a327df7bd6d8465996951842/.semgrep.yml

### What's implemented

The backend implemented here is minimal, but enough to be functional:
`pants lint --only=semgrep ...` in our repo ~identical to running it
externally (except for more caching). Here's some features:

- Automatic configuration finding, by looking for files called
`.semgrep.yml`, `.semgrep.yaml`, or in a `.semgrep/` directory
- Flexible 'targetless'-ness (kinda): 
- semgrep is invoked against _all_ targets with a `source`, relying on
Semgrep's own file walking to apply the right rules to the right files
within each sandbox.
- Multiple/scoped configuration: 
- Each target is linted using all the configurations in its ancestor
directories (`a/b/c/some_file.py` will be linted using all of the rules
configured in `//`, `a/`, `a/b/` and `a/b/c/`, if available).
  - Targets are partitioned by the sets of rules that apply to them.
- Ignore support: `.semgrepignore` (at the top level) is obeyed (NB.
there's no `skip_semgrep` per-target field).
- `--semgrep-force` for disabling caching: Semgrep supports downloading
and running [external rules](https://semgrep.dev/explore) (e.g. `semgrep
-c p/python` includes a bunch of [Python-specific
rules](https://semgrep.dev/p/python)). However, these are mutable and
cannot be pinned, so caching results may be misleading. With this flag,
`pants lint --only=semgrep --semgrep-force --semgrep-args='-c=p/python'
::` will do a whole fresh run, as desired.

### Limitations/future work

There's various bits of potential improvements that I already know
about, but I think that this backend is useful even without them, and
that they can be iterated on.

1. Rules can have auto-fix-ups configured, and applied via `semgrep scan
--autofix ...`, which would fit neatly into `pants fix ...`. I started
writing this before #18430 was fixed, so this doesn't support that.
2. More precise identification of which files need to be run against
which rules, to increase cache reuse. Currently, every single file
that's a descendent of a directory with a `semgrep_rule_source()` target
is put into a semgrep partition/sandbox, but this could be tightened by
parsing/doing dependency inference on the actual rules:
- by language: if a directory only has `python` rules, there's no need
to run it against Go/PHP/yaml/... files in that directory.
- by path: rules can define inclusion/exclusion rules based on file
paths (and similarly the global ignores implied by a `.semgrepignore`
file)
3. This doesn't include support for testing rules:
https://semgrep.dev/docs/writing-rules/testing-rules/ (this may require
rule YAML files to have explicit targets)
4. Ignoring/disabling semgrep for specific files/targets works okay, but
is a bit awkward:
- currently this rule makes sure to include any `.semgrepignore` files,
but only the one in the root is used (Semgrep doesn't support nested
`.semgrepignore` files:
https://github.com/returntocorp/semgrep/issues/5669)
- it'd be good to have a `skip_semgrep` field similar to `skip_pylint`
etc., but this would need to be added to every target with a source
field, some how
4. This appears to be significantly slower than running semgrep outside
of pants. For instance, the table below shows that the time for running
within pants (13 invocations of semgrep linting subsets of the repo) is
significantly slower than running once on all files outside of pants.
Part of this is set-up time of semgrep (each invocation takes at least
0.6s just to start semgrep), but that doesn't seem to explain it all.
However, semgrep is fast enough that I think this is fine for now, and
the finer-grained caching will recover this time for very large repos.

   | method           | real | CPU  |
   |------------------|------|------|
   | pants*           | 4s?  | 24s? |
   | standalone `-j1` | 2.5s | 2.5s |
   | standalone `-j8` | 1.6s | 4.4s |

(* The pants timing is estimated by checking `--stats-log -ldebug` logs:
I'm using `PANTS_SOURCE=.... pants ...` which doesn't use `pantsd`, and
thus using the raw timings would be unfairly penalising pants set-up
time that would disappear.)

Depending on where this PR goes, I'll file follow-up issues for some of
these.