Version the release asset per-platform PEXs (Cherry-pick of #19683) (#19722)
(Admittedly we should've done this out of the gate)

This change makes it so the PEX we upload to our GitHub Release includes
the Pants version in it as well.

This should help with users sitting behind firewalls needing to
customize the URL for PEXs (by replacing the GitHub parts of the pex
URL). Previously the only way to version the PEXs is behind versioned
directories, since the filename is the same across versions. Now, you
can dump all your Pants PEXs in one directory.

This also has the added benefit of allowing users to click to download
multiple versions in the web UI and not blast an older version. And
lastly, mirrors the wheel names better.

Before: `pants.cp39-linux_x86_64.pex`
After: `pants.2.18.0a0-cp39-linux_x86_64.pex`

- I'll backport this change to the existing 2.18.x releases (by making
additional uploads), so the scie-pants code can stay simple
- Accompanying `scie-pants` change:
https://github.com/pantsbuild/scie-pants/pull/260
- Once that is released, I'll update `MINIMUM_SCIE_PANTS_VERSION` here
to list the new version
- I likely _will not_ cherry-pick this to 2.17.x unless we are already
going to make another 2.17.x release

Co-authored-by: Josh Cannon <joshdcannon@gmail.com>