Improve dep inference versioning scheme (#19790)
https://github.com/pantsbuild/pants/pull/19293 needs to be cherry-picked
back to 2.17.x and 2.18.x, but the version that should be specified
isn't clear.
We could copy `0.0.2`, but if any change landed on `main` but not
cherry-picked backwards that version is incorrect.

So, I'm changing the versioning scheme to embed the
latest-major-changes-branch into the version which forces cherry-picking
to fail and the human should bump the destination branch's patch number

---------

Co-authored-by: Stu Hood <stuhood@gmail.com>