Add test for nested no-infer-dep in Rust Python dep inference (#19804)
This closes #19751 by adding some tests for it. We're pretty sure the
real fix was in #19293, but only incidentally, so this PR makes sure
we've got a regression test specifically for #19751.

I've marked the test for cherrypicking so that we confirm that this is
fixed in the earlier releases too.