Tweak `fix` partitioning to closer align to `lint` partitioning (#19796)
https://github.com/pantsbuild/pants/issues/17403#issuecomment-1682688673
describes a bit of a split between `lint` and `fix` in terms of
batching.

This change fixes #19634 by switching `AbstractFixRequest.files` to
report `self.snapshot.files` which is already de-duplicated. (Note that
the elements in the case of `lint` were the filepaths directly
translated from addresses. Maybe it's worth de-duplicating at a higher
layer, but :man_shrugging: )

This change also fixes #17403 (which really is a duplicate of #15069
manifest in new and exciting ways) by de-duplicating the `fix`-specific
batching strategy. It didn't sit right to me that in the reproduction
repo we did 1 batches of 2 files for `fix`, but 1 batch of 4 files for
`lint`.