Mark the pytype type checker as experimental (Cherry-pick of #19849) (#19851)
Support for the pytype type checker was added in #19440, for 2.18, and
was described as inspired by pyright. The pyright backend is still
considered experimental, and, to me, it seems a bit unlucky to land
support as immediately stable, without giving at least one major release
for experimentation and bugs, for what I imagine is a relatively
complicated tool.

(I don't know anything about pytype, so this is mostly me just pattern
matching. I'm happy to be convinced otherwise and have this continue to
be released in 2.18 as stable.)

Co-authored-by: Huon Wilson <huon@exoflare.io>