Add additional directories with register.py to pants bin deps (#19848)
I compared the output of `find . -name register.py` and the dependencies
of the `src/python/pants/bin:plugins` target: this identified several
backends that seemingly weren't being packaged into the Pants binary,
and thus not available for users, similar to #19836.

- `pants.backend.experimental.openapi.codegen.java`
- `pants.backend.experimental.python.framework.django`
- `pants.backend.experimental.swift`
- `pants.backend.python.providers.experimental.pyenv.custom_install`

There were also two that have been upgraded from experimental to stable,
that were depended upon via the old/deprecated experimental name, but
not the stable name:

- `pants.backend.python.lint.autoflake`
- `pants.backend.python.lint.pyupgrade`

This doesn't try to solve the problem of keeping this up to date. I can
think of two approaches (definitely for a separate PR):

- #17301, with a test that reads the `BUILD` file and compares the
dependencies to a `find` for `register.py` files
- a target that exposes the output of `pants peek ::` as a file, with a
test that reads that file and does the appropriate validation