Port `pantsd` fingerprinting to Rust, and validate from native client (Cherry-pick of #19833) (#19867)
As described in #19775, the native client does not detect a difference
in Pants version between the client and server. This is because the
native client does not implement fingerprinting (which I missed in
review!).

This change moves the implementation of `pantsd`'s fingerprint to Rust,
and then validates the fingerprint when the native client reads
`pantsd`'s metadata to decide whether it can handle the connection (or
else whether it should fall back to the Python implementation).

Fixes #19775.