Introduce by-name `@rule` call syntax (#19797)
As discussed on #18905, and sketched on #19730, this change introduces a
call-by-name syntax for `@rules`.

```python
@rule
async def b(i: int) -> B: ...

@rule
def c() -> C: ...

@rule
async def a() -> A:
    _ = await b(**implicitly(int(1)))
    _ = await c()
    return A()
```

As demonstrated in the tests, direct calls to `@rules` are translated
into an `await Call`, which is currently implemented using `Get`. That
means that (currently), the actual `@rule` name which is specified is
not used in graph solving. @benjyw's work on #19755 (and further work on
#19730) will allow us to pin rule graph solving to the called `@rule`
name.

Positional arguments to rules are not yet supported (since the new
arguments would need to be accounted for in rule graph solving, and at
runtime), and will be followup work.