Move .pids under .pants.d by default (#19911)
We also make the workdir a subdir of `.pants.d`, so that it's easy to 
nuke it without nuking the pids. These are just changes to default
option values. It is still possible to explicitly set those options so
whatever the user wants.

The structure under `.pants.d` is opaque to end users and not part
of the API, so we have the freedom to do this without a deprecation.

Note that we do *not* move `.pants.d/bsp` under `.pants.d/workdir`.
For whatever reason, `bsp` uses that hardcoded directory, rather
than consulting the pants_workdir option. 

So the end result of this change is that `.pants.d` is no longer the
same thing as the workdir, but a looser "directory Pants writes stuff
under, and that you can safely nuke". There are now three things that
go under that dir: pids, workdir and bsp.