Remove references to PyPI/`pantsbuild.pants` wheels in docs (#19924)
This updates our docs to refer to the new release process, using GitHub
Releases, instead of PyPI. The `pantsbuild.pants` package becomes
somewhat of an implementation detail too, so I've adjusted these too.

Also, the docs previously referred to the universal PEX we published to
GitHub Releases which we no longer do (#19450), and thus references to
it are removed too.

---------

Co-authored-by: Eric Arellano <14852634+Eric-Arellano@users.noreply.github.com>
Co-authored-by: Andreas Stenius <andreas.stenius@imanage.com>