add support for mTLS in remote cache usage (Cherry-pick of #19887) (#19960)
This PR adds support for mTLS when using remote caching. The API is
fairly similar to what we've already have, but I've taken the liberty of
doing some type cleanups. I also did some refactoring of the tls setup
code to move some earliers to parse-time rather than conversion time,
and simplify a bit.

I also found a bug with the `alpn` config, which isn't actually valid --
hyper just barfs when that is defined. Not sure how no-one else is
hitting that, but it is what it is. Maybe a version/release issue. Comes
from here:
https://github.com/rustls/hyper-rustls/blob/163b3f539a497ae9c4fa65f55a8133234ef33eb3/src/connector/builder.rs#L46-L51.