Bump `console_ui` dependencies and simplify code (#19931)
Dove into this for some other tasks, and it was quite large and hard to
navigate due to handling the indicatif/prodash branching inline. As I
needed to bump indicatif and console for what I originally wanted to do,
I took the liberty of cleaning it up a bit while at it.

* Bump `console` to `0.15.7`
* Bump `indicatif` to `0.17.7`
* Remove custom `lib.path` for `console_ui` crate
* Split `console_ui.rs` into `lib.rs` for the public parts, then
`instance.rs` and `instance/indicatif.rs` + `instance/prodash.rs` for
supported flavors
* Removed a TODO at (now) line 110 of `indicatif.rs` about removing
`pbar.tick()` after upgrading `indicatif` beyond 0.17.0. There's
`ProgressBar::enable_steady_tick` but that spawns one thread per bar.
That is... A bit of noise, and led to some desync over time (spinners
weren't at same state).
* cc @stuhood -- it wasn't super clear what you expected, but removing
the `pbar.tick()` still freezes the spinner
* `ConsoleUI::with_console_ui_disabled` did the `f.await` in both
branches; so simplified that a bit.
* Updated TODO at now line 41 in `indicatif.rs` that it there's no
`WeakMultiProgress`, so it's not likely a good fit to do that change to
avoid potential ownership issues.