Fix algorithm for gathering Go build requests with coverage. (#20030)
The previous algorithm did not check if a package had already been
traversed, which leads to exponential blowup of the queue.

Go doesn't support dependency cycles, so the previous algorithm
would converge eventually, but before it did so the queue could get
to sizes that were infinite in practice. This happened in a real-world
case.