Optionally retry failed tests (#19760)
This MR adds the ability to rerun Tests. The implementation adds a field
on Process for the attempt number. This uniquifies the Process instance,
which means that the Runtime Graph will not deduplicate it and will run
it again. A wrapper RunProcessWithRetries is provided to rerun a normal
process a number of attempts. `TestResult` is modified to have a tuple
of `FallibleProcessResult`s to contain the retries. The output of the
test goal then aggregates these to create a "succeeded after k attempts"
message, like:
```
✓ src/python/pants/core/goals/testomatic_pass_test.py:batch1 succeeded in 0.56s.
+ src/python/pants/core/goals/testomatic_test.py:batch0 succeeded after 4 attempts in 0.53s.
```
The implementation requires each backend to opt into retries of the test
process. This MR only implements it for the pytest backend.
Documentation describing RunProcessWithRetries and how to use it to
retry tests is added.

The number of retries can be set with the `[test].attempts_default`
config setting. This is currently global, but we could extend the
machinery to act like the timeout field.

This implementation retries whole batches, and does not rebatch only
failed targets. It also doesn't aggregate retried process metadata
(currently not easy to get total run time, for example).