remove cyclic dependency betweeen `fs` and `testutil` crates (#20088)
As described in https://github.com/pantsbuild/pants/issues/20086, there is currently a cyclic dependency between the `fs` and `testutil` crates due to `fs` using a method from `testutil` in its tests and `testutil` using types from `fs` in its public interface.

While `cargo` tolerates this cyclic dependency because it was a "dev" dependency, rust-analyzer (used by VSCode) does not tolerate it and fails to load the Pants crates.

Since the dependency in one direction is only a single method, break the dependency by vendoring the `testutil::make_file` method into the `fs` crate. With that done, rust-analyzer is able to continue.