Add 'cattrs' to 'cattrs' default module mapping (#20093)
Since version [1.10.0](https://catt.rs/en/v22.2.0/history.html#id25),
`cattrs` supports imports from both `cattr` or `cattrs` (following
similar changes in `attrs`, which are already reflected in the default
module mapping). We have had to add `cattrs` to our local
`module_mapping` because it is not included in the default. This change
adds it.