Merge "byte store" and "action cache" provider options (#20115)
This replaces uses of `RemoteCacheProviderOptions` with
`RemoteStoreOptions` (previously `RemoteOptions`): they're very similar,
and only separate because they were previously in very different parts
of the codebase, but this has changed with #19958.

This makes a few additional changes to make the merging smoother:
- removes the "capability cell" sharing optimisation, which enabled
sharing of the `GetCapabilitiesRequest` gRPC request result between the
CAS and the remote execution providers, if both were being used with the
same address. I thought that optimising one-time (on start-up) requests
from 2 to 1 wasn't worth the complexity of plumbing it through these
more generic options, but can restore it if it is.
- renames various fields

This does meant that there's a few options that are seemingly irrelevant
to the action cache providers, e.g. "batch API size limit" is
meaningless. I think this is fine: I think it's fundamental to
supporting multiple providers that there'll be a grab-bag of options,
which some providers support and others do not.

The commits are individually reviewable.

(This is one point of #19902.)