Restore protoc 3.20.1 support on apple silicon (#20118)
As pointed out by @huonw in
https://github.com/pantsbuild/pants/pull/20107#issuecomment-1783986672

Removing support for tool versions on mac because they don't have x86
binaries was premature and can cause regressions for users who are
running rosetta. We've recently merged a few changes having to do with
trying to make arm binaries available to mac users:

- https://github.com/pantsbuild/pants/pull/19945 (shellcheck) This one
is fine because it did not drop support for any versions.
- https://github.com/pantsbuild/pants/pull/20058 (protoc). This dropped
support for the default version on apple silicon. This is an error,
because the version seems to work fine with Rosetta. We revert the
version-dropping here.
- https://github.com/pantsbuild/pants/pull/19940 (coursier). This is
very tricky. We dropped support for some old versions, but made new
versions available and bumped the default version. To support both
native-arm and old versions via rosetta is tricky. Discussed further in
https://github.com/pantsbuild/pants/pull/19940/files#r1375420734

For now we revert the coursier fix as discussed in
https://github.com/pantsbuild/pants/pull/19940#discussion_r1375501101 in
a separate PR: #20120

A fix-forward for coursier is available at
https://github.com/pantsbuild/pants/pull/20121

Co-authored-by: Huon Wilson <huon@exoflare.io>